+++ changelog.md
@@ -1,3 +1,10 @@
+# 0.2.0.0 (Sep. 12, 2015)
+
+* Implement NonDetEff
+* Separate Cut/Coroutine out from Internals
+  * Partial implementation: won't compile yet
+* Extract remaining examples from Internal comments
+
 # 0.1.1.0 (Sep. 12, 2015)
 
 * Warnings clean up